refactor(delivery): SlackId 매핑 책임을 값객체 내부로 이동하고 model 패키지로 구조 변경
